% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bioc2019PanCancerStudy.R
\docType{package}
\name{Bioc2019PanCancerStudy}
\alias{Bioc2019PanCancerStudy}
\alias{Bioc2019PanCancerStudy-package}
\title{BioC2019 Pan-Cancer Workshop}
\description{
In this workshop we will show the capability of TCGAbiolinks and Moonlight,
to integrate multi -omics data from different consortium and to reproduce the
six immune subtypes from TCGA PanCancer and how features (Immune Subtypes,
Oncogenic Processes, Driver Genes and Stemness) can be used by the end user to
expand their understating of their own un-published data.
}
