% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_Bioc2019PanCancerStudy.R
\docType{data}
\name{BRCARnaseqSE}
\alias{BRCARnaseqSE}
\title{A SummarizedExperiment object with 20 TCGA BRCA samples obtained from TCGAbiolinks}
\format{A SummarizedExperiment with 21022 genes and 20 samples}
\usage{
BRCARnaseqSE
}
\description{
A SummarizedExperiment object with 20 TCGA BRCA samples obtained from TCGAbiolinks
}
\keyword{internal}
