% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_Bioc2019PanCancerStudy.R
\docType{data}
\name{ImmuneMW}
\alias{ImmuneMW}
\title{A matrix with Immune subtypes trained on TCGA's dataset}
\format{A matrix with 13742 genes and 6 immune subtypes}
\usage{
ImmuneMW
}
\description{
A matrix with Immune subtypes trained on TCGA's dataset
}
\keyword{internal}
