% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_Bioc2019PanCancerStudy.R
\docType{data}
\name{PCBC_stemSig}
\alias{PCBC_stemSig}
\title{A numeric vector with Stemness signature trained on PCBC's dataset}
\format{A numeric vector with 12955 genes}
\usage{
PCBC_stemSig
}
\description{
A numeric vector with Stemness signature trained on PCBC's dataset
}
\keyword{internal}
