% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_Bioc2019PanCancerStudy.R
\docType{data}
\name{TableDataTypes}
\alias{TableDataTypes}
\title{A data.frame for TCGAbiolinks to falitate multi -omics data retrieval}
\format{A data.frame with 7 rows (datatypes) and 4 columns (parameters)}
\usage{
TableDataTypes
}
\description{
A data.frame for TCGAbiolinks to falitate multi -omics data retrieval
}
\keyword{internal}
